Lab Session 1 Jupyter
I try to complete the assignment in Jupyter but of course some problem are expected, especially regarding with the machine design library. Basically in the library they already put some formula to convert as well as to calculate but all of that become impractical if we don't know the symbol used for calculation as well as the formula and the necessary info to make the formula works.
What do you think should we do next? is it possible to write our own code based on our necessity?