Code for shaft and key
The calculation works well in this file but I've still stumbled on issue with the "display()" line command.
Also, I come on issue with the end calculation result where I want to put "round()" line command but because the mechdesign repository is referred then the outcome is not as I intended, it has number with a lot of digits.
Any how, I really have no idea how I can use this github things. e.g. I cannot delete the old files which is irrelevant and of course is not working.